refactor: separate Workspace and WorkspacePainter in two files
